all login test working except resetPassword as it needs token. Refactored initial test code
