+++ backend/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "test": "mocha ./test/test-server.js",
+    "test": "mocha ./test/test-server.js --exit",
     "start": "node server.js",
     "server": "nodemon server.js",
     "legacy": "nodemon -L server.js"
